Primera version beta de la aplicacion

+++ src/main/java/es/laboticademar/webstore/controllers/ProductController.java
@@ -1,6 +1,6 @@
 package es.laboticademar.webstore.controllers;
 
-import java.util.List;
+import java.math.BigDecimal;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
@@ -9,10 +9,8 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 
-import es.laboticademar.webstore.entities.Usuario;
-import es.laboticademar.webstore.services.DestacadoService;
-import es.laboticademar.webstore.services.ProductService;
-import es.laboticademar.webstore.services.UsuarioService;
+import es.laboticademar.webstore.services.interfaces.DestacadoService;
+import es.laboticademar.webstore.services.interfaces.ProductService;
 
 
 @Controller
@@ -25,12 +23,11 @@ public class ProductController {
     private DestacadoService destacadoService;
 
     @GetMapping("/{id}")
-    public String goProduct(@PathVariable("id") Integer id, Model model) {
+    public String goProduct(@PathVariable("id") BigDecimal id, Model model) {
 
         model.addAttribute("destacados", destacadoService.getAllDestacados());
         model.addAttribute("producto", productService.getProductoById(id));
 
         return "product/product";
     }
-    
 }
